refactor(scripts): extract buildAliasMap into shared util
The IUCN link-id builder (next commit) needs the same label->binomial
alias map that build-species-info.js uses. Move it to scripts/lib/aliases.js
so both builders import from one place.
