+++ scripts/build-species-info.js
@@ -23,12 +23,8 @@ import {
   parseGbifIucn,
   parseWikipediaSummary
 } from './build-species-info.lib.js'
+import { buildAliasMap } from './lib/aliases.js'
 import dictionary from '../src/shared/commonNames/dictionary.json' with { type: 'json' }
-import extras from '../src/shared/commonNames/extras.json' with { type: 'json' }
-import speciesnetSource from '../src/shared/commonNames/sources/speciesnet.json' with { type: 'json' }
-import deepfauneSource from '../src/shared/commonNames/sources/deepfaune.json' with { type: 'json' }
-import manasSource from '../src/shared/commonNames/sources/manas.json' with { type: 'json' }
-import { normalizeScientificName } from '../src/shared/commonNames/normalize.js'
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
 const ROOT = path.resolve(__dirname, '..')
@@ -188,27 +184,6 @@ function loadExisting() {
   } catch {
     return {}
   }
-}
-
-/**
- * Build a label → scientificName alias map from every source that emits both
- * fields. Some models (DeepFaune, Manas, plus our own extras) ship a snake_case
- * label alongside the canonical binomial — without this map, only the binomial
- * key gets enriched and a UI lookup by label sees no IUCN/Wikipedia badge.
- */
-function buildAliasMap() {
-  const aliases = new Map()
-  const sources = [speciesnetSource, deepfauneSource, manasSource, extras]
-  for (const src of sources) {
-    for (const entry of src.entries || []) {
-      if (!entry.scientificName || !entry.label) continue
-      const sci = normalizeScientificName(entry.scientificName)
-      const label = normalizeScientificName(entry.label)
-      if (!sci || !label || sci === label) continue
-      aliases.set(label, sci)
-    }
-  }
-  return aliases
 }
 
 /**
